Update CountryController.cs
Add validation

+++ WEB_API_Task_WebApplication/Controllers/CountryController.cs
@@ -33,6 +33,18 @@ namespace WEB_API_Task_WebApplication.Controllers
         [HttpPost]
         public List<Country> Create(Country item)
         {
+        
+            if (item == null)
+            {
+                ModelState.AddModelError("", "No data for add country");
+                return BadRequest(ModelState);
+            }
+        
+             if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+               
             _countryRepository.Create(item);
 
             return _countryRepository.Entities;
@@ -41,7 +53,18 @@ namespace WEB_API_Task_WebApplication.Controllers
         [Route("Delete")]
         [HttpPost]
         public List<Country>Delete(int Id)
-        {
+        {       
+             if (Id == null)
+            {
+                ModelState.AddModelError("", "No data for deleting country");
+                return BadRequest(ModelState);
+            }
+        
+             if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+        
             _countryRepository.Delete(Id);
 
             return _countryRepository.Entities;
@@ -50,7 +73,18 @@ namespace WEB_API_Task_WebApplication.Controllers
         [Route("Update")]
         [HttpPost]
         public List<Country> Update(Country item) 
-        {
+        {       
+            if (item == null)
+            {
+                ModelState.AddModelError("", "No data for updating country");
+                return BadRequest(ModelState);
+            }
+        
+            if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+            
             _countryRepository.Update(item);
 
             return _countryRepository.Entities;
